Modify Entities Item, Loan and Shelf. Add new migration to DB.
Change Shelf Primaty Key (Id) to Compound key (Id, OwnerId).
Change Id variables type from ulong to int.
Change order Fluent API in ApplicationDbContext.

+++ library.BLL/Entities/Loan.cs
@@ -9,8 +9,8 @@ namespace Library.BLL.Entities
     public class Loan
     {
         [Key]
-        public ulong Id { get; set; }
-        public ulong ItemId { get; set; }
+        public int Id { get; set; }
+        public int ItemId { get; set; }
         [ForeignKey("ItemId")]
         public Item Item { get; set; }
